Removed a smattering of compiler warnings

+++ RecastDemo/Source/ChunkyTriMesh.cpp
@@ -118,12 +118,12 @@ static void subdivide(BoundsItem* items, int nitems, int imin, int imax, int tri
 		if (axis == 0)
 		{
 			// Sort along x-axis
-			qsort(items+imin, inum, sizeof(BoundsItem), compareItemX);
+			qsort(items+imin, static_cast<size_t>(inum), sizeof(BoundsItem), compareItemX);
 		}
 		else if (axis == 1)
 		{
 			// Sort along y-axis
-			qsort(items+imin, inum, sizeof(BoundsItem), compareItemY);
+			qsort(items+imin, static_cast<size_t>(inum), sizeof(BoundsItem), compareItemY);
 		}
 		
 		int isplit = imin+inum/2;
